Allow running commands in the proton environment
This adds the possibility to run custom commands inside the proton environment. This is useful to prevent the command to be run in the system wine environment and use the proton environment instead.

While running a command without protontricks wine updates the prefix while proton downgrades it at the next start.

Using protontricks with the command prevents this behavior and no unnecessary upgrades or downgrades are being done.