Set correct LD_LIBRARY_PATH for new Steam Runtime
The newer Steam Runtime appears to work with the existing
LD_LIBRARY_PATH approach as long as the host glibc is new enough. This
means that older Linux distros may have trouble with the new Steam
Runtime unless running the binaries inside a bwrap container, which
isn't done for the time being.

Fixes #72