Use bwrap with newer Steam Runtime installations
Use the provided `run` script to launch Proton executables using Steam
Runtime containers. A different proxy script is used to ensure that
containers are not launched recursively and to ensure LD_LIBRARY_PATH
can be updated inside the container.

This should match the actual Proton behavior better, as well as allow
compatibility with Linux distros that ship older versions of glibc.

Fixes #72