Remove /var from bwrap mounts
Don't mount /var when using bwrap containerization since it might
cause a conflict with /var/run and /run.

Chances are we won't need the mount anyway: the main purpose is to allow
access to any files Winetricks generates, as well as any files the user
might want to access that are likely under /home or one of the mounts
under /mnt or /run/media.