Add `protontricks-launch` command
Add `protontricks-launch` command to launch Windows executables more
easily. For example, the shortest form to launch an EXE named
`ModTool.exe` is:

$ protontricks-launch ModTool.exe

If the app ID is not provided, the user is prompted to select the Steam
app to select the correct Wine prefix. `--appid` can also be provided to
select a specific app:

$ protontricks-launch --appid 12300 ModTool.exe