Display error dialog when launched from desktop
Display pre-determined error or uncaught exception using an error dialog
when launched from desktop. Prior to this, Protontricks would fail to
show any kind of feedback that an error had occurred if Protontricks had
been launched from desktop.

All log messages generated by Protontricks are also included in the
error dialog. This is achieved by registering an additional log handler
that logs all messages to a temporary file. This file is then used when
an error occurs and the error dialog is displayed.