Add YAD as new GUI provider
Add YAD as new GUI provider alongside Zenity. If both are installed,
prefer YAD.

Zenity has a few long-standing bugs that are unlikely to be fixed due to
the project's inactivity. YAD appears to be more active, and it already
has some features that would be useful for GUI improvements (eg. adding
icons to the app selection dialog).

The oldest YAD version that appears to be shipped on modern Linux
distros appears to be v0.40, so try to target it for now.
If it becomes unfeasible, we could implement a version check for YAD
before falling back to Zenity.