Release 1.6.0
Additional notes for package maintainers:

YAD is now the preferred GUI provider. Packages should install YAD when
available instead of relying on Zenity.

Two `desktop` files are provided under `src/protontricks/data`, which
should be installed as part of the package.
`protontricks-desktop-install` script is provided as a fallback for
environments where this is not possible (eg. pipx).
In most cases the script is unnecessary and should be removed as part of
the packaging process.