Fix duplicate app entries
Steam now lists the main installation path itself in the
`libraryfolders.vdf`. This path can also have a different path but
resolve to one of the existing paths due to one of them being a symlink.
This results in Protontricks finding duplicate appmanifest files unless
the duplicate library paths are removed.

Fixes #118