Fix Steam detection for multiple Steam installs
When both Flatpak and non-Flatpak versions of Steam are installed for
the same user, Steam installation detection might return two paths that
correspond to two different installations of Steam.

Ensure that the two paths returned by `find_steam_path` are the same
when Flatpak sandbox is active and Flatpak installation directory is
found.

Refs flathub/com.github.Matoking.protontricks#10