Launch Wine processes in sandbox using launcher
Launch a separate launcher process that allows commands to be executed
inside the sandbox when using bwrap. This allows Wine processes to be
launched inside an existing sandbox without the overhead of launching a
sandbox for each Wine process.

Combined with the keepalive wineserver process, this further improves
the launch time for Wine commands using bwrap.

Refs #120