Fix Proton discovery on Steam Deck
The Steam Deck version of Steam doesn't seem to populate `config.vdf`
unless the user specifically chooses a different Proton version. In this
case, it uses the Proton version specified in the game's Steam Deck
compatibility profile, which can be found in `appcache/appinfo.vdf`.

Also use the stable version of Proton as fallback, and adjust the
priority of `ToolMapping` entries, which now have the lowest priority.

Fixes #141