Print warning when multiple Steam dirs are found
Up until now, Protontricks silently picked the first Steam directory it
found unless `STEAM_DIR` environment variable was specified. This has caused
some confusion when both Flatpak and non-Flatpak versions of Steam, and
Flatpak version of Protontricks are installed, with Protontricks opting
to pick the Flatpak installation in this scenario.

Solve the issue by printing a warning that lists all found Steam
directories, as well as instructing the user to select the specific
directory in case the correct one was not used automatically.

Refs #162