Revert "Launch bwrap launcher using D-Bus instead of sock"
This is required for the Flatpak version of Protontricks, which does not
work without granting full access to D-Bus session bus with the new
implementation. Namely, without full access the `ProcessExited` signal
is not emitted for some reason, causing any Wine process to hang when
bwrap is in use.

Revert to the earlier deprecated `--socket` option. This allows bwrap to
work with the Flatpak version of Protontricks.

Refs #164

This reverts commit 7b333c6fd10da0fb1e7f51ed3701ba01361ea593.