Fix race condition crash in Wine launcher script
The Wine launcher script checks for an existing wineserver process and
copies a few Wine related environment variables from it to ensure all
Wine processes work together. This is important especially if the user
has launched the game using Steam and wants to run other applications
for the game at the same time.

However, since there is a small gap between retrieving the PID of a
wineserver process and retrieving its environment variables, the
wineserver process might have ceased to exist by the time we're
trying to retrieve its environment variables, causing the process to
crash. While this can technically happen in any environment, it seemed
to happen consistently when the Flatpak version of Protontricks is in
use and bwrap is disabled.

Refs #164