Parse appinfo.vdf v28
New version of appinfo.vdf was introduced in Steam beta. The new version
introduces a new SHA1 hash field that encodes the hash for the binary
VDF data, but is otherwise identical to the existing file structure.

Update Protontricks to parse both the new and previous version for the
time being. The code for parsing v27 can be removed once the change is
in stable as well.

Fixes #189