Find Windows shortcut prefix properly
Instead of assuming the Wine prefix for Windows shortcuts exist in the
main Steam installation directory, find it properly using the same
method as the Steam apps.

While the Wine prefix has always been stored in the same directory until
now (or at least on my local machine), this might not be the case in the
future.