Fix locale settings on Steam Deck
The Steam interface on Steam Deck appears to configure locales without
actually generating them (eg. running `locale-gen`). This results in
some applications such as Zenity crashing when the locales can't
actually be loaded.

Add a workaround that detects the locale settings and sets any missing
locales back to `en_US.UTF-8`, as we can probably assume that always
exists. This also means that the workaround is not used if the user has
either generated them manually or if the issue is fixed in the distro.

Fixes #188