Update section on community packages
Remove the manually updated listing; the widget provided by Repology has
far more detail and allows the user to tell which packages are
up-to-date at a glance, making it far more useful.

Also add note on Debian and Ubuntu, as they are a common culprit when it
comes to shipping outdated packages (either Protontricks itself or
Winetricks).

Fixes #197