Check for 'steamapps' in case-insensitive manner
Check for the 'steamapps' directory in a case-insensitive manner.

While 'steamapps' seems to be the default used by Steam,
'SteamApps' has been used in the past and other variations such as
'Steamapps' also seem to exist for one reason or the other.
Steam itself seems to work fine regardless of the exact name as long as
only one directory exists under each library folder.

Fixes #203