Always look for Flatpak Steam installation
Always look for Flatpak Steam regardless of whether we're in a Flatpak
sandbox or not. Protontricks is perfectly capable of handling both
versions of Steam regardless of the execution environment.

Previously we only looked for the Flatpak installation of Steam when
running Protontricks itself in a Flatpak sandbox. This was mostly done
to ensure the (likely) correct Steam installation was chosen
automatically depending on the environment; the assumption was that the
user would likely use Flatpak for both Protontricks and Steam, and the
`STEAM_DIR` environment variable to select a specific installation was
considered too cumbersome for the average user. We've since added a
graphical "select Steam installation" dialog when multiple Steam
installations are found, so it makes no sense to enable Flatpak Steam
detection only for Flatpak Protontricks anymore.