Disable `background-wineserver` by default
Disable `background-wineserver` by default, as it resolves two
common and longstanding issues with typical Protontricks usage when the
`wineserver` process is preserved between commands:

* Graphical applications can crash due to an issue likely related to
  different X11 libraries used inside the Steam Runtime. This issue is
  tracked in #166.
* Console output can become broken. This issue was reported in #223 (and
  likely other GitHub issues as well).

The original reason for implementing the background `wineserver` process
was to shorten `wine` command launch time, as Winetricks could take
nearly a minute to launch in some situations. This seems to be mostly a
non-issue now, with a `wine cmd.exe /c echo %ProgramFiles%` command
taking around 0.7 seconds to launch on my system, and `protontricks
<appid> good` taking around 20 seconds to complete, which is still very
reasonable compared to 10 seconds with `--background-wineserver`.

Fixes #166.
Fixes #223.