Do not use loop to check for launcher
Do not use a loop to sleep until the Steam Runtime
launcher has finished starting up. Instead, create a pipe and pass it to
the underlying `steam-runtime-launcher-service` executable; the
executable will write to it and close it to indicate the launcher has
finished starting up.

After this every subsequent `wine_launch.sh` invocation can assume the
launcher is up instead of checking for the socket in a loop.

This approach was suggested by @smcv on GitHub.

Fixes #231