Add more logging
Add more log statements including multiple DEBUG level log messages.

This should hopefully help with user issues where a single faulty entry
(eg. app manifest) causes a crash. Prior to this, the user was usually
instructed to launch Protontricks in a debugger to determine the exact
entry that causes Protontricks to crash; this is unnecessarily difficult
and time consuming for both the user and the person responding to the
report.

The issue template won't be updated to include the `-vv` flag yet. We
want to release a newer Protontricks version first.