Skip unidentifiable icon files
Pillow cannot parse all image files (usually JPEG). This might be due to
the files not being valid image files in the first place, or Pillow not
being able to parse them. In any case, skip the file and log the error
if the image file cannot be parsed.

Fixes #264
Fixes #268