Add `--cwd-app` flag
Add `--cwd-app` flag to `protontricks` and `protontricks-launch` to set
the working directory to the game's installation directory. Also
document how to maintain previous behavior in the changelog.

This also makes it easier to maintain
backwards compatibility with scripts that assume installation directory
as the working directory. Scripts relying on the working directory
should be infrequent based on searching GitHub repositories for
`protontricks` usage, however.

Fixes #287