Support appinfo.vdf V29
Steam beta introduced a new version of appinfo.vdf with a space-saving
optimization. Field keys are stored in a separate table at the end of
the file, with the actual VDF segments having to be parsed using the
table to map the indices to actual field names.

`vdf` library does not support serializing appinfo.vdf using this
format, at least yet, so just use appinfo.vdf V28 in tests for the time
being. This might need to be fixed in the future once appinfo.vdf V28 is
phased out.

Fixes #304