Vendor patched version of 'vdf'
Vendor patched version of 'vdf' library and use it in case the installed
version does not support the `key_table` argument which we need to
actually parse the `appinfo.vdf` files created by the newest stable
Steam.

This approach is cursed and will probably upset package maintainers,
but it seems like the only reasonable option aside from maintaining a
beta version of Protontricks indefinitely until a new 'vdf' release
appears on PyPI.

PACKAGE MAINTAINERS: This commit can be dropped if the system 'vdf'
already has the `key_table` argument (e.g. it uses the
@Solsticegamestudios fork).