Set arguments explicitly in CLI entrypoints
Set arguments explicitly in CLI entrypoints instead of letting
`argparse` handle it.

This allows us to determine whether any arguments were actually
provided without having to resort to checking `sys.argv`. This is done
more for the sake of tests, which relied on monkeypatching `sys.argv`
which is *very* brittle and broke tests in the Nix build sandbox and
other environments as well.

Fixes #346