Print available Proton versions on missing value
Print available Proton versions when $PROTON_VERSION has been set but
the installation in question cannot be found, as the list of allowed
values is otherwise hard to determine.

Also refactor the Proton app discovery in CLI entrypoint, as it was
already duplicated and needed to be updated to print the new error
message.