Fix app icon discovery
Previously we could assume the app icon path for Steam apps. Starting
with newer Steam client, this is no longer the case and the icon path
might be located under either of the two paths:

* `<steam_path>/appcache/librarycache/<appid>/<40 hex chars>.jpg` (new)
* `<steam_path/appcache/librarycache/<appid>_icon.jpg`

This also means we have to perform additional IO for each Steam app to
detect the icon location; this is done in `SteamApp.from_appmanifest`.