Set additional Proton env vars
Set various environment variables automatically to ensure DXVK,
vkd3d-proton, nvapi and GStreamer are properly configured. This should
improve compatibility in scenarios in which an application launched
through Steam worked properly, but didn't work when launched through
Protontricks.

The environment variables are set fairly conservatively: env var is only
set if the Proton installation appears to support it and if the user
hasn't already set it themselves.

The environment variables are usually set by the `proton` launcher
script provided by Proton installations and used by Steam. However, it
can't be used for a variety of reasons:

* The command-line interface is not stable and can change from version
  to version
* While written in Python, it isn't suitable for programmatic usage due
  to relying on global variables