Add `get_vdf_steam_shortcuts`
Add function to retrieve a list of shortcuts as they're stored in the
Steam client's `shortcuts.vdf` file. This will be needed because we will
be reading the file in two places:

* Once on every Protontricks run to deserialize them to `SteamApp`
  instances
* And again to retrieve the shortcut's executable location to retrieve
  the game's app launch parameters. This will be implemented in a
  follow-up commit.

The `shortcut_id` variable was also removed in the process, as nothing
was actually done with it despite it being cast into an integer.