Make library folder discovery case-insensitive
Steam will go out of its way to perform certain operations
case-insensitively. For example, when a Steam library folder has been
added, Steam will automatically discover it even if it is renamed to ALL
CAPS. The configuration will not be updated to use the new name,
however, so we have to perform this search ourselves as well.

Adjust Steam library folder discovery to account for this.

Refs #433