Change default Proton versions
Previously we used stable Proton as the default Proton version in case
no other configuration is found. This was changed to Proton
Experimental in Steam client at some point.

The default itself is hardcoded in one of the Steam client's shared
libraries, meaning Protontricks cannot determine it programmatically.
Instead, use Proton Experimental as the new default and fall back to the
old default of stable Proton if Proton Experimental is not available.

Fixes #435