Fix silenced IO errors during library discovery
The exception handler used here was meant to return an empty list of
library folders if the VDF configuration file is empty. However, it also
included the actual parsing process as well, causing any unrelated IO
errors to be treated as the configuration file missing and all found
library folders to be discarded. No warning message was also logged in
this situation, even if it would make perfect sense to do so.

Fix this issue by moving VDF file and by performing the path resolving
for each Steam library path resolving in its own block.

Fixes #436